backend/requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
python-decouple
djangorestframework-simplejwt
gunicorn
whitenoise
langchain
langchain-google-genai
langchain-community
faiss-cpu
